+++ PlanJ/app/src/main/res/navigation/schedule_nav.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/shedule_nav"
+    app:startDestination="@id/scheduleFragment">
+
+    <fragment
+        android:id="@+id/scheduleFragment"
+        android:name="com.boostcamp.planj.ui.schedule.ScheduleFragment"
+        android:label="ScheduleFragment" />
+</navigation>
